+++ src/extension.ts
@@ -3,10 +3,7 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
 import { EnterpriseFileDecorationProvider } from "./providers/enterpriseFileDecorationProvider";
-import {
-  EnterpriseTreeDataProvider,
-  TreeEnterpriseItem,
-} from "./providers/enterpriseTreeDataProvider";
+import { EnterpriseTreeDataProvider, TreeEnterpriseItem, } from "./providers/enterpriseTreeDataProvider";
 import { EnterpriseService } from "./services/enterpriseService";
 import { EnterpriseTextDocumentContentProvider } from "./providers/enterpriseTextContentProvider";
 
@@ -17,19 +14,16 @@ export async function activate(context: vscode.ExtensionContext) {
   let url: string | undefined = config.get("url");
   let reloadConfig = false;
 
-  const rootPath =
-    vscode.workspace.workspaceFolders &&
-    vscode.workspace.workspaceFolders.length > 0
-      ? vscode.workspace.workspaceFolders[0].uri.fsPath
-      : undefined;
+  // get the root path of the workspace
+  const rootPath = vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0 ? vscode.workspace.workspaceFolders[0].uri.fsPath : undefined;
 
+  // ensure a workspace folder is open
   if (!rootPath) {
-    vscode.window.showErrorMessage(
-      "STARLIMS: Working folder not found, open a workspace folder an try again."
-    );
+    vscode.window.showErrorMessage("STARLIMS: Working folder not found, open a workspace folder and try again.");
     return;
   }
-  // ensure extension settings are defined and prompt for values if not
+
+  // ensure STARLIMS URL is defined and prompt for value if not
   if (!url) {
     url = await vscode.window.showInputBox({
       title: "Configure STARLIMS",
@@ -40,33 +34,30 @@ export async function activate(context: vscode.ExtensionContext) {
       await config.update("url", url, false);
       reloadConfig = true;
     } else {
-      vscode.window.showErrorMessage(
-        "Please configure STARLIMS URL in extension settings."
-      );
+      vscode.window.showErrorMessage("Please configure STARLIMS URL in extension settings.");
       return;
     }
   }
 
+  // ensure user and password are defined and prompt for values if not
   if (!user) {
     user = await vscode.window.showInputBox({
       title: "Configure STARLIMS",
-      prompt: "Enter STARLIMS username",
+      prompt: "Enter STARLIMS Username",
       ignoreFocusOut: true,
     });
     if (user) {
       await config.update("user", user.toUpperCase(), false);
       reloadConfig = true;
     } else {
-      vscode.window.showErrorMessage(
-        "Please configure STARLIMS user in extension settings."
-      );
+      vscode.window.showErrorMessage("Please configure STARLIMS user in extension settings.");
     }
   }
 
   if (!password) {
     password = await vscode.window.showInputBox({
       title: "Configure STARLIMS",
-      prompt: `Enter password for STARLIMS user '${user}'`,
+      prompt: `Enter Password for STARLIMS User '${user}'`,
       password: true,
       ignoreFocusOut: true,
     });
@@ -74,52 +65,61 @@ export async function activate(context: vscode.ExtensionContext) {
       await config.update("password", password, false);
       reloadConfig = true;
     } else {
-      vscode.window.showErrorMessage(
-        "Please configure STARLIMS password in extension settings."
-      );
+      vscode.window.showErrorMessage("Please configure STARLIMS password in extension settings.");
     }
   }
 
+  // reload the configuration if it was updated
   if (reloadConfig) {
     config = vscode.workspace.getConfiguration("STARLIMS");
   }
 
+  // create the enterprise service
   const enterpriseService = new EnterpriseService(config);
 
-  // register a text content provider to viewing remote code items. it responds to the starlims:/ URI
-  const enterpriseTextContentProvider =
-    new EnterpriseTextDocumentContentProvider(enterpriseService);
+  // create the output channel for the extension
+  const outputChannel = vscode.window.createOutputChannel("STARLIMS");
 
-  context.subscriptions.push(
-    vscode.workspace.registerTextDocumentContentProvider(
-      "starlims",
-      enterpriseTextContentProvider
-    )
-  );
+  // register a text content provider to viewing remote code items. it responds to the starlims:/ URI
+  const enterpriseTextContentProvider = new EnterpriseTextDocumentContentProvider(enterpriseService);
+  context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider("starlims", enterpriseTextContentProvider));
 
   // register a custom tree data provider for the STARLIMS enterprise designer explorer
   const enterpriseProvider = new EnterpriseTreeDataProvider(enterpriseService);
   vscode.window.registerTreeDataProvider("STARLIMS", enterpriseProvider);
 
+  // execute the STARLIMS.Save command when a document is saved
+  vscode.workspace.onDidSaveTextDocument(async (document: vscode.TextDocument) => {
+    // check if the document is in the configured workspace
+    if (document.uri.fsPath.startsWith(rootPath)) {
+      vscode.commands.executeCommand("STARLIMS.Save", document.uri);
+    }
+  });
+  
   // hook into tree events for loading code items
-  vscode.commands.registerCommand(
-    "STARLIMS.selectEnterpriseItem",
+  vscode.commands.registerCommand("STARLIMS.selectEnterpriseItem",
     async (item: TreeEnterpriseItem) => {
-      // open only leaf nodes
+      // open leaf nodes only
       if (item.collapsibleState !== vscode.TreeItemCollapsibleState.None) {
         return;
       }
 
-      const fileExtension =
-        item.language !== undefined &&
-        item.language !== "" &&
-        item.language !== "N/A"
-          ? item.language.toLowerCase()
-          : "txt";
+      // check if the item is already open, switch the tab if it is
+      const openDocument = vscode.workspace.textDocuments.find((doc) => doc.uri.fsPath === item.filePath);
+      if (openDocument) {
+        await vscode.window.showTextDocument(openDocument);
+        return;
+      }
 
-      const uri = vscode.Uri.parse(`starlims://${item.uri}.${fileExtension}`);
-      const doc = await vscode.workspace.openTextDocument(uri); // calls back into the provider
-      await vscode.window.showTextDocument(doc, { preview: false });
+      // get local copy of the item
+      const localFilePath = await enterpriseService.getLocalCopy(item.uri, rootPath);
+
+      // open the file locally
+      if (localFilePath) {
+        item.filePath = localFilePath;
+        let uri: vscode.Uri = vscode.Uri.file(localFilePath);
+        vscode.window.showTextDocument(uri);
+      }
     }
   );
 
@@ -128,19 +128,13 @@ export async function activate(context: vscode.ExtensionContext) {
   vscode.window.registerFileDecorationProvider(fileDecorationProvider);
 
   // this command activates the extension
-  vscode.commands.registerCommand("STARLIMS.Connect", () => {});
+  vscode.commands.registerCommand("STARLIMS.Connect", () => { });
 
-  // registers the GetLocal version command handler.
-  vscode.commands.registerCommand(
-    "STARLIMS.GetLocal",
+  // registers the GetLocal command handler
+  vscode.commands.registerCommand("STARLIMS.GetLocal",
     async (item: TreeEnterpriseItem | any) => {
-      const localFilePath = await enterpriseService.getLocalCopy(
-        item.uri ||
-          (item.path
-            ? item.path.slice(0, item.path.lastIndexOf("."))
-            : undefined),
-        rootPath
-      );
+      // get local copy of the item
+      const localFilePath = await enterpriseService.getLocalCopy(item.uri || (item.path ? item.path.slice(0, item.path.lastIndexOf(".")) : undefined), rootPath);
       if (localFilePath) {
         let uri: vscode.Uri = vscode.Uri.file(localFilePath);
         vscode.window.showTextDocument(uri);
@@ -148,20 +142,11 @@ export async function activate(context: vscode.ExtensionContext) {
     }
   );
 
-  const outputChannel = vscode.window.createOutputChannel("STARLIMS");
-
-  // registers the GetLocal version command handler.
-  vscode.commands.registerCommand(
-    "STARLIMS.RunScript",
+  // registers the RunScript command handler
+  vscode.commands.registerCommand("STARLIMS.RunScript",
     async (item: TreeEnterpriseItem | any) => {
-      const uri =
-        item.uri ||
-        (item.path
-          ? item.path.slice(0, item.path.lastIndexOf("."))
-          : undefined);
-      outputChannel.appendLine(
-        `${new Date().toLocaleString()} Executing remote script at URI: ${uri}`
-      );
+      const uri = (item.path ? item.path.slice(0, item.path.lastIndexOf(".")) : undefined);
+      outputChannel.appendLine(`${new Date().toLocaleString()} Executing remote script at URI: ${uri}`);
       const result = await enterpriseService.runScript(uri);
       if (result) {
         outputChannel.appendLine(result);
@@ -171,10 +156,9 @@ export async function activate(context: vscode.ExtensionContext) {
   );
 
   // registers the remote compare command
-  vscode.commands.registerCommand(
-    "STARLIMS.Compare",
+  vscode.commands.registerCommand("STARLIMS.Compare",
     async (uri: vscode.Uri) => {
-      // is command executed on the file tree
+      // command executed on the file tree
       let localUri = uri;
       if (!localUri) {
         // if not, compare with the open document
@@ -192,9 +176,7 @@ export async function activate(context: vscode.ExtensionContext) {
           let remoteUri = vscode.Uri.parse(`starlims://${remotePath}`);
           vscode.commands.executeCommand("vscode.diff", remoteUri, localUri);
         } else {
-          vscode.window.showErrorMessage(
-            "STARLIMS: Working folder not found, open a workspace folder an try again."
-          );
+          vscode.window.showErrorMessage("STARLIMS: Working folder not found, open a workspace folder an try again.");
         }
       }
     }
@@ -204,7 +186,14 @@ export async function activate(context: vscode.ExtensionContext) {
   vscode.commands.registerCommand(
     "STARLIMS.Checkout",
     async (item: TreeEnterpriseItem) => {
-      await enterpriseService.checkout(item.uri);
+      let bSuccess = await enterpriseService.CheckOut(item.uri);
+      if (bSuccess) {
+        item.checkedOutBy = user;
+        enterpriseProvider.refresh();
+
+        // execute getlocal command
+        vscode.commands.executeCommand("STARLIMS.GetLocal", item);
+      }
     }
   );
 
@@ -212,27 +201,45 @@ export async function activate(context: vscode.ExtensionContext) {
   vscode.commands.registerCommand(
     "STARLIMS.Checkin",
     async (item: TreeEnterpriseItem) => {
-      let checkinReason: string =
-        (await vscode.window.showInputBox({
+      let checkinReason: string = (await vscode.window.showInputBox({
           prompt: "Enter checkin reason",
           ignoreFocusOut: true,
-        })) || "";
+        })) || "Checked in from VSCode";
 
-      await enterpriseService.checkin(item.uri, checkinReason);
+      let bSuccess = await enterpriseService.CheckIn(item.uri, checkinReason);
+      if (bSuccess) {
+        item.checkedOutBy = "";
+        enterpriseProvider.refresh();
+      }
     }
   );
 
+  // registers the refresh command
   vscode.commands.registerCommand(
-    "STARLIMS.refresh",
+    "STARLIMS.Refresh",
     async (item: TreeEnterpriseItem) => {
       await enterpriseProvider.refresh();
     }
   );
 
-  vscode.window.showInformationMessage(
-    `Connected to STARLIMS on ${config.url}.`
-  );
+  // register the save file command
+  vscode.commands.registerCommand("STARLIMS.Save",
+    async (item: TreeEnterpriseItem) => {
+      const editor = vscode.window.activeTextEditor;
+      if (editor) {
+        var localUri = editor.document.uri;
+        if (localUri) {
+          if (vscode.workspace.workspaceFolders !== undefined) {
+            const workspaceFolderPath = vscode.workspace.workspaceFolders[0].uri.path;
+            let remotePath = localUri.path.slice(workspaceFolderPath.length);
+            enterpriseService.saveEnterpriseItemCode(remotePath, editor.document.getText());
+          }
+        }
+      }
+    });
+
+  vscode.window.showInformationMessage(`Connected to STARLIMS on ${config.url}.`);
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }